split katana60 into revisions
