+++ keyboards/rominronin/katana60/rev1/config.h
@@ -25,7 +25,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define PRODUCT_ID      0x6060
 #define DEVICE_VER      0x0001
 #define MANUFACTURER    RominRonin CandyKeys
-#define PRODUCT         Katana60
+#define PRODUCT         Katana60 rev1
 
 /* key matrix size */
 #define MATRIX_ROWS 5
