+++ keyboards/rominronin/katana60/rev1/info.json
@@ -1,6 +1,6 @@
 {
-  "keyboard_name": "Katana60",
-  "url": "https://candykeys.com/product/katana60-pcb-V2",
+  "keyboard_name": "Katana60 rev1",
+  "url": "",
   "maintainer": "rominronin",
   "width": 15,
   "height": 5,
