+++ keyboards/rominronin/katana60/rev1/readme.md
@@ -1,16 +1,16 @@
-# Katana60
+# Katana60 rev1
 
-![Katana60](https://i.imgur.com/xVkODOu.jpg)
+![Katana60 rev1](https://i.imgur.com/xVkODOu.jpg)
 
 A 60% keyboard with a symmetrical staggered layout, sold by CandyKeys.com.
 
 Keyboard Maintainer: [Baris Tosun](https://github.com/rominronin)  
-Hardware Supported: Katana60 PCB v1
-Hardware Availability: the v1 is no longer available, the v2 can be found at [candykeys.com](https://candykeys.com/product/katana60-pcb-V2)
+Hardware Supported: Katana60 PCB rev1
+Hardware Availability: the rev1 is no longer available, the rev2 can be found at [candykeys.com](https://candykeys.com/product/katana60-pcb-V2)
 
 Make example for this keyboard (after setting up your build environment):
 
-    make rominronin/katana60:default
+    make rominronin/katana60/rev1:default
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
