+++ keyboards/rominronin/katana60/rev1/rev1.c
@@ -13,4 +13,4 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "katana60.h"
+#include "rev1.h"
