+++ keyboards/rominronin/katana60/rev2/info.json
@@ -1,7 +1,7 @@
 {
-   "keyboard_name":"",
-   "url":"",
-   "maintainer":"qmk",
+   "keyboard_name":"Katana60 rev1",
+   "url":"https://candykeys.com/product/katana60-pcb-V2",
+   "maintainer":"rominronin",
    "width":19,
    "height":15.5,
    "layouts":{
