+++ keyboards/rominronin/katana60/rev2/readme.md
@@ -1,16 +1,16 @@
-# katana60_rev2
+# Katana60 rev2
 
-![katana60_rev2](https://i.imgur.com/3hBsSvB.jpg)
+![Katana60 rev2](https://i.imgur.com/3hBsSvB.jpg)
 
 A 60% keyboard with a symmetrical staggered layout, sold by CandyKeys.com.
 
 Keyboard Maintainer: [Baris Tosun](https://github.com/rominronin)  
-Hardware Supported: Katana60 PCB v2/rev2/revision 2
+Hardware Supported: Katana60 PCB rev2
 Hardware Availability: [candykeys.com](https://candykeys.com/product/katana60-pcb-V2)
 
 Make example for this keyboard (after setting up your build environment):
 
-    make rominronin/katana60_rev2:default
+    make rominronin/katana60/rev2:default
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
