+++ keyboards/rominronin/katana60/rev2/rev2.c
@@ -13,4 +13,4 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "katana60_rev2.h"
+#include "rev2.h"
